+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/from_request_bodies/media_type_shape.rs
@@ -1,0 +1,15 @@
+#[derive(Clone, Debug)]
+pub(super) enum MediaTypeShape {
+    ApplicationJson,
+    Unsupported(String),
+}
+
+impl MediaTypeShape {
+    pub fn new<A: Into<String>>(key: A) -> MediaTypeShape {
+        let key = key.into();
+        match key.as_str() {
+            "application/json" => MediaTypeShape::ApplicationJson,
+            x => MediaTypeShape::Unsupported(x.into()),
+        }
+    }
+}
