+++ libs/openapi-types/src/v3_0/media_type.rs
@@ -8,6 +8,15 @@ impl MediaTypeKey {
     pub fn new<A: Into<String>>(a: A) -> Self {
         Self(a.into())
     }
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl From<MediaTypeKey> for String {
+    fn from(this: MediaTypeKey) -> Self {
+        this.0
+    }
 }
 
 /// rf. https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#media-type-object
